Another go, trying to make more "functional"
Moved testing logic into simple function which takes the testing
functions, and also added curried function specifically for a test
of score < 0.78 and average < 0.70.

Moved state map variable into its own actor, so at least the only
way to access and change it is through GetScores and SetScores
messages to the CurrentScores actor.  Still having trouble trying
to puzzle out how to get rid of this, or if it's even possible
(leaning towards it's an unfortunate necessity as the past scores
must simply be kept for the next score request).

The actor still just sets an entire list and returns an entire list
depending on the key.  This way the caller must still set up the
list which will be updated.  The actor only isolates the keeping
of the state (i.e. the "side-effects" part).

Removed exceptions and instead use Try[T] to turn the exception
into a Try object which can be pattern matched, and an appropriate
response sent.

Chained up the futures, so that the AddressFraudProbability score
will get chained with the retrieval of the score list which will
get chained with the actual pass/fail function call, which will
have its future returned.  Then this future can be awaited (with
the requisite 5 second timeout) to get the final result.

Any errors/exceptions or issues in the above chain will abort
further future chains and just return a failed future all the
way through, which will show up as a Try.Failure object when
the Await is called (either from a  timeout or another error).